Change url to kebab case and query es by uid
